fixes #3, #5
Added functionality for a setting to enable capture on startup.
If we're capturing on startup and we're connected to a valid DB,
the extension will not ask the user to choose a new DB on startup.

Added import/export queries to the extension. At some point I need
to decide if I want to store the queries in the settings or not.

Also changed the storage location of settings and queries so they
persist between Burp instances. They're not stored in the project
anymore. Had a failed attempt to add a setting to store them in
the project or in the Burp instance, but it did not work correctly.